requirements.txt
streamlit
torch
torchvision
librosa
matplotlib
numpy
pandas
opencv-python
pillow
scipy